Remove useless one shot code
